api_test/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
firebase-admin==7.0.0
pyrebase4==4.6.0
requests==2.32.4
python-dotenv==1.1.1
azure-monitor-opentelemetry==1.6.10
opentelemetry-instrumentation-fastapi==0.49b2
redis==5.0.1
psycopg2-binary==2.9.9
asyncpg==0.30.0
